Changes required to authorize the notification endpoint.
